fix(#399): strip the doc-page mount harness at manifest generation, not just serve-time
The mount-isolation harness (mountIdPrefix / $props.id() / derived ids) that
the component doc page injects to scope element ids across its double-mount was
only stripped at the playground /example-src/ serve route. The manifest
generator embedded the raw harnessed source into each shipped
<component>.examples.json 'code' field — the snippet npm consumers are told to
copy verbatim — leaking meaningless, uncopyable plumbing into 28 examples.

- Move stripExampleHarness to packages/components/scripts/lib/ so the generator
  (which shapes the published artifact) and the playground server share one
  fail-closed source of truth. scripts/ is excluded from the tarball, so there
  is no consumer surface-area cost; playground reaches back via the established
  ../../components/scripts/... relative-import pattern.
- Wire it into extractExampleFile after buildCodeField; a fail-closed throw
  becomes a hard extraction error that blocks components:generate (and
  publishing) rather than shipping half-stripped code.
- Regenerate all 28 manifests + components.json (zero drift; zero harness leak).
- Split the test suite along the new ownership line: transform fixture/unit
  tests live with the function in components; the route-glue and full-corpus
  strip+svelte-compile sweep stay in playground, importing across the boundary.
- Move FEATURED_EXAMPLES + source-shape predicates into a plain
  featured-examples.ts module so two tests share the registry without one test
  importing another's exports.

Closes #399
