+++ packages/components/scripts/lib/strip-example-harness.ts
