+++ packages/components/src/components/autocomplete/autocomplete.examples.json
@@ -7,7 +7,7 @@
       "id": "async",
       "title": "Async source",
       "description": "Remote-style source with a loading row and stale-request cancellation.",
-      "code": "<script lang=\"ts\">\n  import Autocomplete from '@lostgradient/cinder/autocomplete';\n\n  let { mountIdPrefix }: { mountIdPrefix?: string } = $props();\n  const uid = $props.id();\n  let fieldId = $derived(`${mountIdPrefix ?? uid}-field`);\n\n  const cities = [\n    { value: 'Austin, Texas', description: 'United States' },\n    { value: 'Auckland, New Zealand', description: 'New Zealand' },\n    { value: 'Aurora, Colorado', description: 'United States' },\n    { value: 'Berlin, Germany', description: 'Germany' },\n    { value: 'Bristol, England', description: 'United Kingdom' },\n  ];\n\n  let value = $state('');\n\n  async function suggestionSource(query: string, context: { signal: AbortSignal }) {\n    await new Promise<void>((resolve, reject) => {\n      const timeout = setTimeout(resolve, 350);\n      context.signal.addEventListener(\n        'abort',\n        () => {\n          clearTimeout(timeout);\n          reject(new DOMException('Aborted', 'AbortError'));\n        },\n        { once: true },\n      );\n    });\n\n    return cities.filter((city) => city.value.toLowerCase().includes(query.toLowerCase()));\n  }\n</script>\n\n<Autocomplete\n  id={fieldId}\n  label=\"Destination city\"\n  placeholder=\"Type a city\"\n  {suggestionSource}\n  bind:value\n/>\n"
+      "code": "<script lang=\"ts\">\n  import Autocomplete from '@lostgradient/cinder/autocomplete';\n\n  const cities = [\n    { value: 'Austin, Texas', description: 'United States' },\n    { value: 'Auckland, New Zealand', description: 'New Zealand' },\n    { value: 'Aurora, Colorado', description: 'United States' },\n    { value: 'Berlin, Germany', description: 'Germany' },\n    { value: 'Bristol, England', description: 'United Kingdom' },\n  ];\n\n  let value = $state('');\n\n  async function suggestionSource(query: string, context: { signal: AbortSignal }) {\n    await new Promise<void>((resolve, reject) => {\n      const timeout = setTimeout(resolve, 350);\n      context.signal.addEventListener(\n        'abort',\n        () => {\n          clearTimeout(timeout);\n          reject(new DOMException('Aborted', 'AbortError'));\n        },\n        { once: true },\n      );\n    });\n\n    return cities.filter((city) => city.value.toLowerCase().includes(query.toLowerCase()));\n  }\n</script>\n\n<Autocomplete\n  id=\"field\"\n  label=\"Destination city\"\n  placeholder=\"Type a city\"\n  {suggestionSource}\n  bind:value\n/>\n"
     },
     {
       "id": "basic",
