+++ packages/components/src/components/chat/chat.examples.json
@@ -7,7 +7,7 @@
       "id": "basic",
       "title": "Basic chat",
       "description": "A conversation with editable input and markdown rendering.",
-      "code": "<script lang=\"ts\">\n  import {\n    Chat,\n    appendAssistantMessage,\n    appendUserMessage,\n    createConversation,\n  } from '@lostgradient/cinder/chat';\n\n  let { mountIdPrefix }: { mountIdPrefix?: string } = $props();\n  const uid = $props.id();\n  let chatId = $derived(`${mountIdPrefix ?? uid}-chat`);\n\n  // `id` below is the conversation snapshot's data key — it never reaches a DOM\n  // element id (Chat derives every rendered id from its `id` prop, scoped above\n  // as `chatId`), so it needs no mount-prefix scoping.\n  let conversation = $state(\n    appendAssistantMessage(\n      appendUserMessage(createConversation({ id: 'basic-chat' }), 'Can you summarize this plan?'),\n      `The plan is ready to implement.\n\n1. Port the component surface.\n2. Add focused demos.\n3. Keep the public API importable by subpath.`,\n    ),\n  );\n</script>\n\n<div style=\"height: 34rem;\">\n  <Chat id={chatId} {conversation} allowAttachments={false} />\n</div>\n"
+      "code": "<script lang=\"ts\">\n  import {\n    Chat,\n    appendAssistantMessage,\n    appendUserMessage,\n    createConversation,\n  } from '@lostgradient/cinder/chat';\n\n  // `id` below is the conversation snapshot's data key — it never reaches a DOM\n  // element id (Chat derives every rendered id from its `id` prop, scoped above\n  // as `chatId`), so it needs no mount-prefix scoping.\n  let conversation = $state(\n    appendAssistantMessage(\n      appendUserMessage(createConversation({ id: 'basic-chat' }), 'Can you summarize this plan?'),\n      `The plan is ready to implement.\n\n1. Port the component surface.\n2. Add focused demos.\n3. Keep the public API importable by subpath.`,\n    ),\n  );\n</script>\n\n<div style=\"height: 34rem;\">\n  <Chat id=\"chat\" {conversation} allowAttachments={false} />\n</div>\n"
     },
     {
       "id": "interactive-harness",
