+++ packages/components/src/components/drawer/drawer.examples.json
@@ -7,7 +7,7 @@
       "id": "basic",
       "title": "Basic drawer",
       "description": "An edge-anchored slide-in panel with controls for side, size, and triggerRef.",
-      "code": "<script lang=\"ts\">\n  import { Button } from '@lostgradient/cinder/button';\n  import { Checkbox } from '@lostgradient/cinder/checkbox';\n  import { Drawer } from '@lostgradient/cinder/drawer';\n  import type { DrawerSide, DrawerSize } from '@lostgradient/cinder/drawer';\n  import { Select } from '@lostgradient/cinder/select';\n\n  let { mountIdPrefix }: { mountIdPrefix?: string } = $props();\n  const uid = $props.id();\n  let sideId = $derived(`${mountIdPrefix ?? uid}-side`);\n  let sizeId = $derived(`${mountIdPrefix ?? uid}-size`);\n  let useTriggerRefId = $derived(`${mountIdPrefix ?? uid}-use-trigger-ref`);\n\n  let open = $state(false);\n  let triggerRef: HTMLElement | null = $state(null);\n  let side = $state<DrawerSide>('right');\n  let size = $state<DrawerSize>('md');\n  let useTriggerRef = $state(true);\n\n  const sideOptions: { value: DrawerSide; label: string }[] = [\n    { value: 'right', label: 'right' },\n    { value: 'left', label: 'left' },\n  ];\n  const sizeOptions: { value: DrawerSize; label: string }[] = [\n    { value: 'sm', label: 'sm' },\n    { value: 'md', label: 'md' },\n    { value: 'lg', label: 'lg' },\n    { value: 'xl', label: 'xl' },\n  ];\n</script>\n\n<div style=\"display: flex; flex-direction: column; gap: 1rem; max-width: 32rem;\">\n  <div style=\"display: flex; flex-wrap: wrap; gap: 1rem; align-items: flex-end;\">\n    <Select id={sideId} bind:value={side} options={sideOptions} label=\"Side\" />\n    <Select id={sizeId} bind:value={size} options={sizeOptions} label=\"Size\" />\n    <Checkbox id={useTriggerRefId} bind:checked={useTriggerRef} label=\"Use triggerRef\" />\n  </div>\n\n  <Button\n    label=\"Open drawer\"\n    onclick={(event) => {\n      triggerRef = useTriggerRef ? event.currentTarget : null;\n      open = true;\n    }}\n  />\n</div>\n\n<Drawer bind:open {side} {size} title=\"Drawer panel\" triggerRef={useTriggerRef ? triggerRef : null}>\n  <p>This is the drawer body. You can put any content here.</p>\n  <p>Current settings: side={side}, size={size}.</p>\n\n  {#snippet footer()}\n    <div style=\"display: flex; gap: 0.5rem;\">\n      <Button variant=\"secondary\" label=\"Cancel\" onclick={() => (open = false)} />\n      <Button label=\"Confirm\" onclick={() => (open = false)} />\n    </div>\n  {/snippet}\n</Drawer>\n"
+      "code": "<script lang=\"ts\">\n  import { Button } from '@lostgradient/cinder/button';\n  import { Checkbox } from '@lostgradient/cinder/checkbox';\n  import { Drawer } from '@lostgradient/cinder/drawer';\n  import type { DrawerSide, DrawerSize } from '@lostgradient/cinder/drawer';\n  import { Select } from '@lostgradient/cinder/select';\n\n  let open = $state(false);\n  let triggerRef: HTMLElement | null = $state(null);\n  let side = $state<DrawerSide>('right');\n  let size = $state<DrawerSize>('md');\n  let useTriggerRef = $state(true);\n\n  const sideOptions: { value: DrawerSide; label: string }[] = [\n    { value: 'right', label: 'right' },\n    { value: 'left', label: 'left' },\n  ];\n  const sizeOptions: { value: DrawerSize; label: string }[] = [\n    { value: 'sm', label: 'sm' },\n    { value: 'md', label: 'md' },\n    { value: 'lg', label: 'lg' },\n    { value: 'xl', label: 'xl' },\n  ];\n</script>\n\n<div style=\"display: flex; flex-direction: column; gap: 1rem; max-width: 32rem;\">\n  <div style=\"display: flex; flex-wrap: wrap; gap: 1rem; align-items: flex-end;\">\n    <Select id=\"side\" bind:value={side} options={sideOptions} label=\"Side\" />\n    <Select id=\"size\" bind:value={size} options={sizeOptions} label=\"Size\" />\n    <Checkbox id=\"use-trigger-ref\" bind:checked={useTriggerRef} label=\"Use triggerRef\" />\n  </div>\n\n  <Button\n    label=\"Open drawer\"\n    onclick={(event) => {\n      triggerRef = useTriggerRef ? event.currentTarget : null;\n      open = true;\n    }}\n  />\n</div>\n\n<Drawer bind:open {side} {size} title=\"Drawer panel\" triggerRef={useTriggerRef ? triggerRef : null}>\n  <p>This is the drawer body. You can put any content here.</p>\n  <p>Current settings: side={side}, size={size}.</p>\n\n  {#snippet footer()}\n    <div style=\"display: flex; gap: 0.5rem;\">\n      <Button variant=\"secondary\" label=\"Cancel\" onclick={() => (open = false)} />\n      <Button label=\"Confirm\" onclick={() => (open = false)} />\n    </div>\n  {/snippet}\n</Drawer>\n"
     },
     {
       "id": "overflowing",
