+++ packages/components/src/components/form-section/form-section.examples.json
@@ -7,7 +7,7 @@
       "id": "account-settings",
       "title": "Account settings panel",
       "description": "FormSection groups related account and privacy controls under a headed section. Text fields use FormField + Input; on/off settings use Toggle standalone (it renders its own label).",
-      "code": "<script lang=\"ts\">\n  import { FormField } from '@lostgradient/cinder/form-field';\n  import { FormSection } from '@lostgradient/cinder/form-section';\n  import { Input } from '@lostgradient/cinder/input';\n  import { Toggle } from '@lostgradient/cinder/toggle';\n\n  let { mountIdPrefix }: { mountIdPrefix?: string } = $props();\n  const uid = $props.id();\n  let displayNameId = $derived(`${mountIdPrefix ?? uid}-display-name`);\n  let accountEmailId = $derived(`${mountIdPrefix ?? uid}-account-email`);\n  let publicProfileId = $derived(`${mountIdPrefix ?? uid}-public-profile`);\n  let marketingEmailsId = $derived(`${mountIdPrefix ?? uid}-marketing-emails`);\n\n  let displayName = $state('');\n  let email = $state('');\n  let publicProfile = $state(true);\n  let marketingEmails = $state(false);\n</script>\n\n<div style=\"display: flex; flex-direction: column; gap: 2rem;\">\n  <FormSection\n    heading=\"Account\"\n    headingLevel={2}\n    description=\"Update your display name and email address.\"\n    columns={2}\n  >\n    <FormField id={displayNameId} label=\"Display name\">\n      <Input id={displayNameId} bind:value={displayName} placeholder=\"Jane Smith\" />\n    </FormField>\n    <FormField id={accountEmailId} label=\"Email address\">\n      <Input id={accountEmailId} bind:value={email} type=\"email\" placeholder=\"jane@example.com\" />\n    </FormField>\n  </FormSection>\n\n  <FormSection\n    heading=\"Privacy\"\n    headingLevel={2}\n    description=\"Control who can see your profile and what communications you receive.\"\n  >\n    <Toggle id={publicProfileId} bind:checked={publicProfile} label=\"Public profile\" />\n    <Toggle id={marketingEmailsId} bind:checked={marketingEmails} label=\"Marketing emails\" />\n  </FormSection>\n</div>\n"
+      "code": "<script lang=\"ts\">\n  import { FormField } from '@lostgradient/cinder/form-field';\n  import { FormSection } from '@lostgradient/cinder/form-section';\n  import { Input } from '@lostgradient/cinder/input';\n  import { Toggle } from '@lostgradient/cinder/toggle';\n\n  let displayName = $state('');\n  let email = $state('');\n  let publicProfile = $state(true);\n  let marketingEmails = $state(false);\n</script>\n\n<div style=\"display: flex; flex-direction: column; gap: 2rem;\">\n  <FormSection\n    heading=\"Account\"\n    headingLevel={2}\n    description=\"Update your display name and email address.\"\n    columns={2}\n  >\n    <FormField id=\"display-name\" label=\"Display name\">\n      <Input id=\"display-name\" bind:value={displayName} placeholder=\"Jane Smith\" />\n    </FormField>\n    <FormField id=\"account-email\" label=\"Email address\">\n      <Input id=\"account-email\" bind:value={email} type=\"email\" placeholder=\"jane@example.com\" />\n    </FormField>\n  </FormSection>\n\n  <FormSection\n    heading=\"Privacy\"\n    headingLevel={2}\n    description=\"Control who can see your profile and what communications you receive.\"\n  >\n    <Toggle id=\"public-profile\" bind:checked={publicProfile} label=\"Public profile\" />\n    <Toggle id=\"marketing-emails\" bind:checked={marketingEmails} label=\"Marketing emails\" />\n  </FormSection>\n</div>\n"
     },
     {
       "id": "fieldset-legend",
