+++ packages/components/src/components/markdown-editor/markdown-editor.examples.json
@@ -7,7 +7,7 @@
       "id": "basic",
       "title": "Basic markdown editor",
       "description": "A Milkdown-backed editor with toolbar and source mode toggle.",
-      "code": "<script lang=\"ts\">\n  import { MarkdownEditor } from '@lostgradient/cinder/markdown-editor';\n\n  let { mountIdPrefix }: { mountIdPrefix?: string } = $props();\n  const uid = $props.id();\n  let editorId = $derived(`${mountIdPrefix ?? uid}-editor`);\n\n  let value = $state(`# Release notes\n\nThis editor supports **rich markdown editing**, toolbar actions, and a source mode.\n\n- Write in markdown\n- Switch between rendered and source modes\n- Keep the value bindable`);\n</script>\n\n<div style=\"display: grid; gap: 0.75rem;\">\n  <MarkdownEditor id={editorId} bind:value showToolbar showModeToggle />\n  <p style=\"margin: 0; color: var(--cinder-text-muted);\">\n    Characters: {value.length}\n  </p>\n</div>\n"
+      "code": "<script lang=\"ts\">\n  import { MarkdownEditor } from '@lostgradient/cinder/markdown-editor';\n\n  let value = $state(`# Release notes\n\nThis editor supports **rich markdown editing**, toolbar actions, and a source mode.\n\n- Write in markdown\n- Switch between rendered and source modes\n- Keep the value bindable`);\n</script>\n\n<div style=\"display: grid; gap: 0.75rem;\">\n  <MarkdownEditor id=\"editor\" bind:value showToolbar showModeToggle />\n  <p style=\"margin: 0; color: var(--cinder-text-muted);\">\n    Characters: {value.length}\n  </p>\n</div>\n"
     }
   ]
 }
