+++ packages/components/src/components/modal/modal.examples.json
@@ -7,7 +7,7 @@
       "id": "basic",
       "title": "Basic modal",
       "description": "A modal used as a generic content shell — in this case an invite form with structured fields.",
-      "code": "<script lang=\"ts\">\n  import { Button } from '@lostgradient/cinder/button';\n  import { Input } from '@lostgradient/cinder/input';\n  import { Modal } from '@lostgradient/cinder/modal';\n\n  let { mountIdPrefix }: { mountIdPrefix?: string } = $props();\n  const uid = $props.id();\n  let nameId = $derived(`${mountIdPrefix ?? uid}-name`);\n  let emailId = $derived(`${mountIdPrefix ?? uid}-email`);\n\n  let open = $state(false);\n  let triggerRef: HTMLElement | null = $state(null);\n  let email = $state('');\n  let name = $state('');\n\n  function handleSend() {\n    open = false;\n  }\n</script>\n\n<Button\n  label=\"Invite teammate\"\n  onclick={(event) => {\n    triggerRef = event.currentTarget;\n    open = true;\n  }}\n/>\n\n<Modal bind:open title=\"Invite teammate\" {triggerRef}>\n  <div style=\"display: grid; gap: 1rem;\">\n    <Input id={nameId} label=\"Full name\" placeholder=\"Alex Kim\" bind:value={name} autofocus />\n    <Input\n      id={emailId}\n      label=\"Email address\"\n      type=\"email\"\n      placeholder=\"alex@example.com\"\n      bind:value={email}\n    />\n  </div>\n\n  {#snippet footer()}\n    <div style=\"display: flex; gap: 0.5rem; justify-content: flex-end;\">\n      <Button variant=\"secondary\" label=\"Cancel\" onclick={() => (open = false)} />\n      <Button variant=\"primary\" label=\"Send invite\" onclick={handleSend} />\n    </div>\n  {/snippet}\n</Modal>\n"
+      "code": "<script lang=\"ts\">\n  import { Button } from '@lostgradient/cinder/button';\n  import { Input } from '@lostgradient/cinder/input';\n  import { Modal } from '@lostgradient/cinder/modal';\n\n  let open = $state(false);\n  let triggerRef: HTMLElement | null = $state(null);\n  let email = $state('');\n  let name = $state('');\n\n  function handleSend() {\n    open = false;\n  }\n</script>\n\n<Button\n  label=\"Invite teammate\"\n  onclick={(event) => {\n    triggerRef = event.currentTarget;\n    open = true;\n  }}\n/>\n\n<Modal bind:open title=\"Invite teammate\" {triggerRef}>\n  <div style=\"display: grid; gap: 1rem;\">\n    <Input id=\"name\" label=\"Full name\" placeholder=\"Alex Kim\" bind:value={name} autofocus />\n    <Input\n      id=\"email\"\n      label=\"Email address\"\n      type=\"email\"\n      placeholder=\"alex@example.com\"\n      bind:value={email}\n    />\n  </div>\n\n  {#snippet footer()}\n    <div style=\"display: flex; gap: 0.5rem; justify-content: flex-end;\">\n      <Button variant=\"secondary\" label=\"Cancel\" onclick={() => (open = false)} />\n      <Button variant=\"primary\" label=\"Send invite\" onclick={handleSend} />\n    </div>\n  {/snippet}\n</Modal>\n"
     },
     {
       "id": "overflowing",
