+++ packages/components/src/components/review-editor/review-editor.examples.json
@@ -7,7 +7,7 @@
       "id": "basic",
       "title": "Basic review editor",
       "description": "Editable markdown with source, rendered, and diff views.",
-      "code": "<script lang=\"ts\">\n  import { ReviewEditor } from '@lostgradient/cinder/review-editor';\n\n  let { mountIdPrefix }: { mountIdPrefix?: string } = $props();\n  const uid = $props.id();\n  let editorId = $derived(`${mountIdPrefix ?? uid}-review-editor`);\n\n  const original = `# Release Plan\n\nThe first release includes a dashboard, export actions, and inline review.\n\n## Checklist\n\n- Finalize the component API\n- Add playground coverage\n- Verify consumer fixtures`;\n\n  let value = $state(`# Release Plan\n\nThe first release includes a dashboard, export actions, inline review, and comment summaries.\n\n## Checklist\n\n- Finalize the component API\n- Add playground coverage\n- Verify consumer fixtures\n- Document review export behavior`);\n</script>\n\n<div style=\"min-height: 34rem;\">\n  <ReviewEditor\n    id={editorId}\n    {original}\n    bind:value\n    currentUserId=\"steve\"\n    name=\"review\"\n  />\n</div>\n"
+      "code": "<script lang=\"ts\">\n  import { ReviewEditor } from '@lostgradient/cinder/review-editor';\n\n  const original = `# Release Plan\n\nThe first release includes a dashboard, export actions, and inline review.\n\n## Checklist\n\n- Finalize the component API\n- Add playground coverage\n- Verify consumer fixtures`;\n\n  let value = $state(`# Release Plan\n\nThe first release includes a dashboard, export actions, inline review, and comment summaries.\n\n## Checklist\n\n- Finalize the component API\n- Add playground coverage\n- Verify consumer fixtures\n- Document review export behavior`);\n</script>\n\n<div style=\"min-height: 34rem;\">\n  <ReviewEditor id=\"review-editor\" {original} bind:value currentUserId=\"steve\" name=\"review\" />\n</div>\n"
     },
     {
       "id": "with-comments",
