+++ packages/components/src/components/search-field/search-field.examples.json
@@ -7,7 +7,7 @@
       "id": "basic",
       "title": "Clearable search field",
       "description": "Search input with an initial value so the clear button is visible.",
-      "code": "<script lang=\"ts\">\n  import { SearchField } from '@lostgradient/cinder/search-field';\n\n  let { mountIdPrefix }: { mountIdPrefix?: string } = $props();\n  const uid = $props.id();\n  let fieldId = $derived(`${mountIdPrefix ?? uid}-field`);\n</script>\n\n<SearchField\n  id={fieldId}\n  aria-label=\"Search components\"\n  defaultValue=\"cinder\"\n  placeholder=\"Filter by name\"\n/>\n"
+      "code": "<script lang=\"ts\">\n  import { SearchField } from '@lostgradient/cinder/search-field';\n\n</script>\n\n<SearchField\n  id=\"field\"\n  aria-label=\"Search components\"\n  defaultValue=\"cinder\"\n  placeholder=\"Filter by name\"\n/>\n"
     }
   ]
 }
