+++ packages/components/src/components/segmented-control/segmented-control.examples.json
@@ -7,7 +7,7 @@
       "id": "basic",
       "title": "Basic segmented control",
       "description": "A compact radio-group control with keyboard navigation.",
-      "code": "<script lang=\"ts\">\n  import { Segment } from '@lostgradient/cinder/segment';\n  import { SegmentedControl } from '@lostgradient/cinder/segmented-control';\n\n  let { mountIdPrefix }: { mountIdPrefix?: string } = $props();\n  const uid = $props.id();\n  let controlId = $derived(`${mountIdPrefix ?? uid}-playground-view`);\n\n  let value = $state('rendered');\n</script>\n\n<div style=\"display: grid; gap: 0.75rem; justify-items: start;\">\n  <SegmentedControl id={controlId} selectionMode=\"single\" bind:value label=\"Document view\">\n    <Segment value=\"source\">Source</Segment>\n    <Segment value=\"rendered\">Rendered</Segment>\n    <Segment value=\"diff\">Diff</Segment>\n  </SegmentedControl>\n  <p style=\"margin: 0; color: var(--cinder-text-muted);\">Current view: {value}</p>\n</div>\n"
+      "code": "<script lang=\"ts\">\n  import { Segment } from '@lostgradient/cinder/segment';\n  import { SegmentedControl } from '@lostgradient/cinder/segmented-control';\n\n  let value = $state('rendered');\n</script>\n\n<div style=\"display: grid; gap: 0.75rem; justify-items: start;\">\n  <SegmentedControl id=\"playground-view\" selectionMode=\"single\" bind:value label=\"Document view\">\n    <Segment value=\"source\">Source</Segment>\n    <Segment value=\"rendered\">Rendered</Segment>\n    <Segment value=\"diff\">Diff</Segment>\n  </SegmentedControl>\n  <p style=\"margin: 0; color: var(--cinder-text-muted);\">Current view: {value}</p>\n</div>\n"
     },
     {
       "id": "multiple",
