+++ packages/components/src/components/skip-link/skip-link.examples.json
@@ -7,7 +7,7 @@
       "id": "basic",
       "title": "Basic usage",
       "description": "Tab into this example to reveal the skip link, then press Enter to jump focus to the main content area.",
-      "code": "<script lang=\"ts\">\n  import { SkipLink } from '@lostgradient/cinder/skip-link';\n\n  let { mountIdPrefix }: { mountIdPrefix?: string } = $props();\n  const uid = $props.id();\n  let mainId = $derived(`${mountIdPrefix ?? uid}-main`);\n</script>\n\n<div\n  style=\"position: relative; min-height: 80px; border: 1px dashed var(--cinder-border); border-radius: 4px; padding: 1rem;\"\n>\n  <SkipLink target={mainId} />\n\n  <nav aria-label=\"Example navigation\">\n    <a href={`#${mainId}`} style=\"margin-right: 1rem;\">Home</a>\n    <a href={`#${mainId}`} style=\"margin-right: 1rem;\">About</a>\n    <a href={`#${mainId}`}>Contact</a>\n  </nav>\n\n  <main id={mainId} style=\"margin-top: 1rem;\">\n    <p>Main content area. Focus lands here when the skip link is activated.</p>\n  </main>\n</div>\n"
+      "code": "<script lang=\"ts\">\n  import { SkipLink } from '@lostgradient/cinder/skip-link';\n\n</script>\n\n<div\n  style=\"position: relative; min-height: 80px; border: 1px dashed var(--cinder-border); border-radius: 4px; padding: 1rem;\"\n>\n  <SkipLink target=\"main\" />\n\n  <nav aria-label=\"Example navigation\">\n    <a href=\"#main\" style=\"margin-right: 1rem;\">Home</a>\n    <a href=\"#main\" style=\"margin-right: 1rem;\">About</a>\n    <a href=\"#main\">Contact</a>\n  </nav>\n\n  <main id=\"main\" style=\"margin-top: 1rem;\">\n    <p>Main content area. Focus lands here when the skip link is activated.</p>\n  </main>\n</div>\n"
     }
   ]
 }
