+++ packages/components/src/components/tag-input/tag-input.examples.json
@@ -7,7 +7,7 @@
       "id": "basic",
       "title": "Basic tag input",
       "description": "Commit free-form tags with Enter or a comma delimiter.",
-      "code": "<script lang=\"ts\">\n  import { FormField, TagInput } from '@lostgradient/cinder';\n\n  let { mountIdPrefix }: { mountIdPrefix?: string } = $props();\n  const uid = $props.id();\n  let fieldId = $derived(`${mountIdPrefix ?? uid}-field`);\n</script>\n\n<FormField\n  id={fieldId}\n  label=\"Topics\"\n  description=\"Press Enter or comma to turn the current text into a tag.\"\n>\n  <TagInput placeholder=\"Add a topic…\" />\n</FormField>\n"
+      "code": "<script lang=\"ts\">\n  import { FormField, TagInput } from '@lostgradient/cinder';\n\n</script>\n\n<FormField\n  id=\"field\"\n  label=\"Topics\"\n  description=\"Press Enter or comma to turn the current text into a tag.\"\n>\n  <TagInput placeholder=\"Add a topic…\" />\n</FormField>\n"
     },
     {
       "id": "disabled",
