+++ packages/playground/src/strip-example-harness.test.ts
@@ -1,19 +1,21 @@
 /**
- * Tests for {@link stripExampleHarness} — the serve-time strip that removes the
- * doc-page mount-isolation harness (`mountIdPrefix` / `$props.id()`) from the
- * reader-facing "Show code" snippet (#399).
+ * Playground-side tests for the example-harness strip (#399). The transform
+ * itself is owned by `packages/components` (the manifest generator shapes the
+ * shipped `code` field with it); its fixture/unit suite lives there at
+ * `scripts/lib/strip-example-harness.test.ts`. This file covers the two things
+ * that need playground's own surroundings:
  *
- * Three layers:
- *   1. Fixture assertions on exact output for each distinct harness shape.
- *   2. Route integration: {@link exampleSnippetResponse} turns a successful
- *      strip into a 200 and a fail-closed throw into a controlled 500.
- *   3. A parametrized sweep over EVERY real example file that contains the
+ *   1. Route integration: {@link exampleSnippetResponse} turns a successful strip
+ *      into a 200 and a fail-closed throw into a controlled 500.
+ *   2. A parametrized sweep over EVERY real example file that contains the
  *      harness: strip it, assert no harness symbol survives, and compile the
  *      stripped output with `svelte/compiler` to prove it is still valid,
- *      copy-pasteable Svelte (catches any dangling derived-id reference a
- *      narrower string assertion would miss). Its count is cross-checked against
- *      the featured-example registry so a moved/renamed dir can't silently zero
- *      the sweep.
+ *      copy-pasteable Svelte (catches any dangling derived-id reference a narrower
+ *      string assertion would miss). Its count is cross-checked against the
+ *      featured-example registry so a moved/renamed dir can't silently zero the
+ *      sweep. This is also the canary the codex-advisor flagged: a new harness
+ *      shape that the fail-closed parser can't handle would now break
+ *      `components:generate`, so this corpus sweep must stay green.
  */
 
 import { describe, expect, it } from 'bun:test';
@@ -22,140 +24,11 @@ import { basename, dirname, join } from 'node:path';
 
 import { compile } from 'svelte/compiler';
 
-import { FEATURED_EXAMPLES } from './examples/featured-example-mount-id.test.ts';
+import { stripExampleHarness } from '../../components/scripts/lib/strip-example-harness.ts';
+import { FEATURED_EXAMPLES } from './examples/featured-examples.ts';
 import { exampleSnippetResponse } from './playground-server.ts';
-import { stripExampleHarness } from './strip-example-harness.ts';
 
 const EXAMPLES_ROOT = join(import.meta.dir, 'examples');
-
-describe('stripExampleHarness — fixtures', () => {
-  it('strips the harness and rewrites a single suffix-form id (input/basic)', () => {
-    const source = `<script lang="ts">
-  import { Input } from '@lostgradient/cinder/input';
-
-  let { mountIdPrefix }: { mountIdPrefix?: string } = $props();
-  const uid = $props.id();
-  let fieldId = $derived(\`\${mountIdPrefix ?? uid}-field\`);
-
-  let name = $state('');
-</script>
-
-<Input id={fieldId} bind:value={name} label="Full name" placeholder="Jane Smith" />
-`;
-    const stripped = stripExampleHarness(source, 'input/basic');
-    expect(stripped).toBe(`<script lang="ts">
-  import { Input } from '@lostgradient/cinder/input';
-
-  let name = $state('');
-</script>
-
-<Input id="field" bind:value={name} label="Full name" placeholder="Jane Smith" />
-`);
-  });
-
-  it('rewrites multiple suffix-form ids (modal/basic shape)', () => {
-    const source = `<script lang="ts">
-  let { mountIdPrefix }: { mountIdPrefix?: string } = $props();
-  const uid = $props.id();
-  let nameId = $derived(\`\${mountIdPrefix ?? uid}-name\`);
-  let emailId = $derived(\`\${mountIdPrefix ?? uid}-email\`);
-</script>
-
-<Input id={nameId} />
-<Input id={emailId} />
-`;
-    const stripped = stripExampleHarness(source, 'modal/basic');
-    expect(stripped).toContain('<Input id="name" />');
-    expect(stripped).toContain('<Input id="email" />');
-    expect(stripped).not.toContain('mountIdPrefix');
-    expect(stripped).not.toContain('$props.id()');
-  });
-
-  it('rewrites a name= binding (radio-group shape)', () => {
-    const source = `<script lang="ts">
-  let { mountIdPrefix }: { mountIdPrefix?: string } = $props();
-  const uid = $props.id();
-  let groupId = $derived(\`\${mountIdPrefix ?? uid}-plan\`);
-</script>
-
-<RadioGroup name={groupId} id={groupId}>x</RadioGroup>
-`;
-    const stripped = stripExampleHarness(source, 'radio-group/basic');
-    expect(stripped).toContain('<RadioGroup name="plan" id="plan">');
-    expect(stripped).not.toContain('groupId');
-  });
-
-  it('rewrites template interpolation and attribute bindings (skip-link/basic)', () => {
-    const source = `<script lang="ts">
-  let { mountIdPrefix }: { mountIdPrefix?: string } = $props();
-  const uid = $props.id();
-  let mainId = $derived(\`\${mountIdPrefix ?? uid}-main\`);
-</script>
-
-<SkipLink target={mainId} />
-<a href={\`#\${mainId}\`}>Home</a>
-<main id={mainId}>x</main>
-`;
-    const stripped = stripExampleHarness(source, 'skip-link/basic');
-    expect(stripped).toContain('<SkipLink target="main" />');
-    expect(stripped).toContain('<a href="#main">Home</a>');
-    expect(stripped).toContain('<main id="main">');
-    expect(stripped).not.toContain('mainId');
-  });
-
-  it('leaves a harness-free example untouched', () => {
-    const source = `<script lang="ts">
-  import { Button } from '@lostgradient/cinder/button';
-</script>
-
-<Button label="Click me" />
-`;
-    expect(stripExampleHarness(source, 'button/basic')).toBe(source);
-  });
-
-  it('fails closed when a removed identifier survives an unhandled binding', () => {
-    // `data-foo={fieldId}` is not one of the rewritten binding forms, so the
-    // identifier survives — the strip must throw rather than ship broken code.
-    const source = `<script lang="ts">
-  let { mountIdPrefix }: { mountIdPrefix?: string } = $props();
-  const uid = $props.id();
-  let fieldId = $derived(\`\${mountIdPrefix ?? uid}-field\`);
-</script>
-
-<Input data-foo={fieldId} />
-`;
-    expect(() => stripExampleHarness(source, 'input/basic')).toThrow(/still referenced/);
-  });
-
-  it('fails closed on a $-prefixed identifier the word-boundary check would miss', () => {
-    // `$rootId` ends and begins with characters `\b` does not treat as a word
-    // boundary; identifier boundaries must still catch the surviving reference.
-    const source = `<script lang="ts">
-  let { mountIdPrefix }: { mountIdPrefix?: string } = $props();
-  const uid = $props.id();
-  let $rootId = $derived(\`\${mountIdPrefix ?? uid}-root\`);
-</script>
-
-<Thing data-x={$rootId} />
-`;
-    expect(() => stripExampleHarness(source, 'thing/basic')).toThrow(/still referenced/);
-  });
-
-  it('fails closed when an unrecognized derived shape leaves mountIdPrefix behind', () => {
-    // A `const` (not `let`) derived declaration is not matched by DERIVED_ID_LINE,
-    // so the prop/uid lines are removed but `mountIdPrefix ?? uid` survives. The
-    // residual-marker scan must catch this rather than serve half-stripped code.
-    const source = `<script lang="ts">
-  let { mountIdPrefix }: { mountIdPrefix?: string } = $props();
-  const uid = $props.id();
-  const fieldId = $derived(\`\${mountIdPrefix ?? uid}-field\`);
-</script>
-
-<Input id={fieldId} />
-`;
-    expect(() => stripExampleHarness(source, 'input/basic')).toThrow(/harness marker/);
-  });
-});
 
 describe('exampleSnippetResponse — route glue', () => {
   it('returns a 200 text/plain stripped snippet on success', async () => {
